chore: rename `lib/pool` => `lib/core/cmap`
