+++ lib/core/cmap/connection_manager.js
