+++ lib/core/cmap/counter.js
