+++ lib/core/cmap/errors.js
