+++ lib/core/cmap/events.js
