+++ lib/core/cmap/index.js
