+++ lib/core/cmap/pool.js
@@ -2,7 +2,7 @@
 
 const EventEmitter = require('events').EventEmitter;
 const makeCounter = require('./counter').makeCounter;
-const Connection = require('./connection').Connection;
+const Connection = require('./connection').CMAPConnection;
 const WaitQueue = require('./wait_queue').WaitQueue;
 const ConnectionManager = require('./connection_manager').ConnectionManager;
